feat(designer) use OperationInfo and ParameterInfo types
