+++ libs/designer/src/lib/ui/panel/nodeDetailsPanel/tabs/parametersTab/index.tsx
@@ -33,8 +33,15 @@ import type { Settings } from '../../../../settings/settingsection';
 import { ConnectionDisplay } from './connectionDisplay';
 import { IdentitySelector } from './identityselector';
 import { MessageBar, MessageBarType, Spinner, SpinnerSize } from '@fluentui/react';
-import { EditorService, toEditorAndOptions } from '@microsoft/designer-client-services-logic-apps';
-import { DynamicCallStatus, PanelLocation, TokenPicker, TokenPickerButtonLocation, TokenType } from '@microsoft/designer-ui';
+import { EditorService } from '@microsoft/designer-client-services-logic-apps';
+import {
+  DynamicCallStatus,
+  PanelLocation,
+  TokenPicker,
+  TokenPickerButtonLocation,
+  TokenType,
+  toCustomEditorAndOptions,
+} from '@microsoft/designer-ui';
 import type { ChangeState, ParameterInfo, ValueSegment, OutputToken, TokenPickerMode, PanelTabFn } from '@microsoft/designer-ui';
 import type { OperationInfo } from '@microsoft/utils-logic-apps';
 import { equals, getPropertyValue } from '@microsoft/utils-logic-apps';
@@ -422,17 +429,11 @@ export const getEditorAndOptions = (
   variables: Record<string, VariableDeclaration[]>
 ): { editor?: string; editorOptions?: any } => {
   const customEditor = EditorService()?.getEditor({
-    connectorId: operationInfo.connectorId,
-    operationId: operationInfo.operationId,
-    parameterKey: parameter.parameterKey,
-    parameterName: parameter.parameterName,
-    required: parameter.required,
-    schema: parameter.schema,
-    editor: parameter.editor,
-    editorOptions: parameter.editorOptions,
+    operationInfo,
+    parameter,
   });
   if (customEditor) {
-    return toEditorAndOptions(customEditor);
+    return toCustomEditorAndOptions(customEditor);
   }
 
   const { editor, editorOptions } = parameter;
